requirements.txt
Django==2.0
django-filter==1.1.0
django-input-mask==2.0.2
django-localflavor==2.0
django-localflavor-br==1.0
django-material==1.2.0
django-viewflow==1.2.0
olefile==0.44
Pillow==4.3.0
pytz==2017.3
reportlab==3.4.0
six==1.11.0
